[Refactor] seperated serializers to each serializer
- remove dependency with pre-refactored serializers.py
