--- actoon/serializer.py
@@ -1,92 +1,0 @@
-from django.contrib.auth import get_user_model
-from rest_framework import serializers
-
-from actoon.models import Project, Task, Media, Effect, Cut
-
-
-class ProjectSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Project
-        fields = [
-            'name',
-            'description'
-        ]
-
-
-class TaskListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Task
-        fields = [
-            'effect',
-            'cut',
-            'image_properties',
-            'project',
-            'parameters',
-            'created_at'
-        ]
-
-
-class TaskSerializer(serializers.ModelSerializer):
-    effect_name = serializers.CharField(max_length=50)
-
-    class Meta:
-        model = Task
-        fields = [
-            'effect',
-            'effect_name',
-            'cut',
-            'image_properties',
-            'project',
-            'parameters',
-            'created_at'
-        ]
-        read_only_fields = ['effect', 'project', 'created_at']
-
-    def create(self, validated_data):
-        validated_data.pop('effect_name')  # remove the field which is not a member
-        return Task.objects.create(**validated_data)
-
-
-class MediaSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Media
-        fields = ['media_type', 'project', 'file', 'proceeded', 'proceeded_image']
-        read_only_fields = ['project', 'proceeded', 'proceeded_image']
-
-
-class EffectSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Effect
-        fields = ['name', 'required_parameters']
-
-
-class UserSerializer(serializers.ModelSerializer):
-    password = serializers.CharField(write_only=True)
-
-    class Meta:
-        model = get_user_model()
-        fields = ['id', 'username', 'password', 'email']
-        write_only_fields = ['password']
-        read_only_fields = ['id']
-
-    def create(self, validated_data):
-        user = get_user_model().objects.create(
-            username=validated_data['username']
-        )
-        user.set_password(validated_data['password'])
-        user.save()
-
-        return user
-
-
-class CutSerializer(serializers.ModelSerializer):
-    file_upload = serializers.FileField(required=False)
-
-    class Meta:
-        model = Cut
-        fields = ['file', 'type', 'sequence', 'sub_sequence', 'pos_x', 'pos_y', 'file_upload']
-        read_only_fields = ['file', 'type', 'sequence', 'sub_sequence', 'pos_x', 'pos_y']
-
-    def create(self, validated_data):
-        validated_data.pop('file_upload')
-        return Cut.objects.create(**validated_data)
