+++ actoon/serializers/projectserializer.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+
+from actoon.models import Project
+
+
+class ProjectSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Project
+        fields = [
+            'name',
+            'description'
+        ]
