+++ actoon/serializers/userserializer.py
@@ -1,0 +1,21 @@
+from django.contrib.auth import get_user_model
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(write_only=True)
+
+    class Meta:
+        model = get_user_model()
+        fields = ['id', 'username', 'password', 'email']
+        write_only_fields = ['password']
+        read_only_fields = ['id']
+
+    def create(self, validated_data):
+        user = get_user_model().objects.create(
+            username=validated_data['username']
+        )
+        user.set_password(validated_data['password'])
+        user.save()
+
+        return user
